@@ -1,32 +1,19 @@
 import * as React from "react";
 import { Provider } from "react-redux";
 import { store, Playground } from "graphql-playground-react";
-import { Session } from "graphql-playground-react/lib/state/sessions/reducers";
 import { TankaClient, TankaLink } from "@tanka/tanka-graphql-server-link";
-import { LogLevel } from "@aspnet/signalr";
-import { ISettings } from 'graphql-playground-react/lib/types';
 import { RetryLink } from "apollo-link-retry";
 import { ApolloLink } from 'apollo-link';
-import { ClientOptions } from '@tanka/tanka-graphql-server-link/dist/client';
 
-var options: ClientOptions = {
-  connection: {
-    logger: LogLevel.Information
-  },
-  reconnectAttempts: Infinity,
-  reconnectInitialWaitMs: 5000,
-  reconnectAdditionalWaitMs: 2000
-};
-
-var client = new TankaClient("https://localhost:5000/graphql", options);
-var tankaLink = new TankaLink(client);
+const client = new TankaClient('https://localhost:5000/graphql', {});
+const tankaLink = new TankaLink(client);
 
 const link = ApolloLink.from([
   new RetryLink(),
   tankaLink
 ]);
 
-var settings: ISettings  = {
+const settings  = {
   'editor.cursorShape': 'line', // possible values: 'line', 'block', 'underline'
   'editor.fontFamily': `'Source Code Pro', 'Consolas', 'Inconsolata', 'Droid Sans Mono', 'Monaco', monospace`,
   'editor.fontSize': 14,
@@ -44,14 +31,14 @@ var settings: ISettings  = {
   'tracing.hideTracingResponse': true,
 };
 
-class App extends React.Component {
-  createLink = (session: Session) => {
+class Home extends React.Component {
+  createLink = (session) => {
     return {
       link: link
     };
   };
 
-  public render() {
+  render() {
     return (
       <Provider store={store}>
         <Playground createApolloLink={this.createLink} settings={settings} />
@@ -60,4 +47,4 @@ class App extends React.Component {
   }
 }
 
-export default App;
+export default Home;
